requirements.txt
flask==2.3.3
flask-executor==1.0.0
flask-wtf==1.2.1
gunicorn==21.2.0
psycopg2-binary==2.9.9
psycopg2-pool==1.1
pandas==2.0.3
openai==1.3.0
python-dotenv==1.0.0
requests==2.31.0
werkzeug==2.3.7
beautifulsoup4==4.12.2
Pillow==10.1.0
numpy==1.24.3

